refactor(common): split agg into multiple files under new mod
