+++ rust/common/src/vector_op/agg/general_sorted_grouper.rs
@@ -1,0 +1,390 @@
+use crate::array::*;
+use crate::error::{ErrorCode, Result};
+use crate::types::*;
+
+/// `EqGroups` encodes the grouping information in the sort aggregate algorithm.
+///
+/// - `SortedGrouper::split_groups` creates a `EqGroups` from a single column.
+/// - `EqGroups::intersect` combines `EqGroups` from each column into a single one.
+/// - `{SortedGrouper,Aggregator}::update_and_output_with_sorted_groups` needs the
+/// grouping information to perform the grouped aggregation.
+///
+/// Internally, `EqGroups` is encoded as the indices that each new group starts.
+/// Specially, a leading `0` means (the 0-th tuple of) this chunk starts a new
+/// group compared to (the last tuple of) the previous chunk, and there is no leading
+/// `0` when it continues the same group or there is no previous chunk.
+pub struct EqGroups(Vec<usize>);
+impl EqGroups {
+    pub fn get_starting_indices(&self) -> &Vec<usize> {
+        &self.0
+    }
+
+    /// `intersect` combines the grouping information from each column into a single one.
+    /// This is required so that we know `group by c1, c2` with `c1 = [a, a, c, c, d, d]`
+    /// and `c2 = [g, h, h, h, h, h]` actually forms 4 groups: `[(a, g), (a, h), (c, h), (d, h)]`.
+    ///
+    /// Since the internal encoding is a sequence of sorted indices, this is effectively
+    /// merging all sequences into a single one with deduplication. In the example above,
+    /// the `EqGroups` of `c1` is `[2, 4]` and that of `c2` is `[1]`, so the output of
+    /// `intersect` would be `[1, 2, 4]` identifying the new groups starting at these indices.
+    pub fn intersect(columns: &[EqGroups]) -> EqGroups {
+        let mut ret = Vec::new();
+        // Use of BinaryHeap here is not to get a performant implementation but a
+        // concise one. The number of group columns would not be huge.
+        // Storing iterator rather than (ci, idx) in heap actually makes the implementation
+        // more verbose:
+        // https://play.rust-lang.org/?version=stable&mode=debug&edition=2018&gist=1e3b098ee3ef352d5a0cac03b3193799
+        use std::cmp::Reverse;
+        use std::collections::BinaryHeap;
+        let mut heap = BinaryHeap::new();
+        for (ci, column) in columns.iter().enumerate() {
+            if let Some(ri) = column.0.get(0) {
+                heap.push(Reverse((ri, ci, 0)));
+            }
+        }
+        while let Some(Reverse((ri, ci, idx))) = heap.pop() {
+            if let Some(ri_next) = columns[ci].0.get(idx + 1) {
+                heap.push(Reverse((ri_next, ci, idx + 1)));
+            }
+            if ret.last() == Some(ri) {
+                continue;
+            }
+            ret.push(*ri);
+        }
+        EqGroups(ret)
+    }
+}
+
+/// `SortedGrouper` contains the state of a group column in the sort aggregate
+/// algorithm, just like `Aggregator` contains the state of an aggregate column.
+pub trait SortedGrouper: Send + 'static {
+    /// `split_groups` detects the `EqGroups` from the `input` array if appended
+    /// to current state. See the documentation of `EqGroups` to learn more.
+    ///
+    /// This is a `dry-run` and does not update its state yet, because it does not
+    /// have grouping information from all group columns yet.
+    fn split_groups(&self, input: &ArrayImpl) -> Result<EqGroups>;
+
+    /// `update_and_output_with_sorted_groups` updates with each subslice of the
+    /// `input` array according to the `EqGroups`. Finished groups are outputted
+    /// to `builder` immediately along the way. After this call, the internal state
+    /// is about the last group which may continue in the next chunk. It can be
+    /// obtained with `output` when there are no more upstream data.
+    fn update_and_output_with_sorted_groups(
+        &mut self,
+        input: &ArrayImpl,
+        builder: &mut ArrayBuilderImpl,
+        groups: &EqGroups,
+    ) -> Result<()>;
+
+    /// `output` the state to the `builder`. Expected to be called once to obtain
+    /// the last group, when there are no more upstream data.
+    fn output(&self, builder: &mut ArrayBuilderImpl) -> Result<()>;
+}
+pub type BoxedSortedGrouper = Box<dyn SortedGrouper>;
+
+pub fn create_sorted_grouper(input_type: DataTypeKind) -> Result<BoxedSortedGrouper> {
+    match input_type {
+        // DataTypeKind::Int16 => Ok(Box::new(GeneralSortedGrouper::<I16Array>::new())),
+        // DataTypeKind::Int32 => Ok(Box::new(GeneralSortedGrouper::<I32Array>::new())),
+        DataTypeKind::Int32 => Ok(Box::new(GeneralSortedGrouper::<I32Array> {
+            ongoing: false,
+            group_value: None,
+        })),
+        // DataTypeKind::Int64 => Ok(Box::new(GeneralSortedGrouper::new::<I64Array>())),
+        unimpl_input => todo!("unsupported sorted grouper: input={:?}", unimpl_input),
+    }
+}
+
+pub struct GeneralSortedGrouper<T>
+where
+    T: Array,
+    for<'a> T::RefItem<'a>: Eq,
+{
+    // Technically `group_value` is meaningless when `ongoing == false` and this
+    // should be a single `Option<Option<T:OwnedItem>>`. But it actually makes this
+    // simple struct less readable.
+    ongoing: bool,
+    group_value: Option<T::OwnedItem>,
+}
+impl<T> GeneralSortedGrouper<T>
+where
+    T: Array,
+    for<'a> T::RefItem<'a>: Eq,
+{
+    pub fn new(ongoing: bool, group_value: Option<T::OwnedItem>) -> Self {
+        Self {
+            ongoing,
+            group_value,
+        }
+    }
+
+    pub fn split_groups_concrete(&self, input: &T) -> Result<EqGroups> {
+        let mut ret = Vec::new();
+        let mut ongoing = self.ongoing;
+        let mut ongoing_group = self.group_value.as_ref().map(|x| x.as_scalar_ref());
+        for (i, v) in input.iter().enumerate() {
+            if ongoing && ongoing_group == v {
+                continue;
+            }
+            if ongoing {
+                ret.push(i);
+            }
+            ongoing = true;
+            ongoing_group = v;
+        }
+        Ok(EqGroups(ret))
+    }
+    pub fn update_and_output_with_sorted_groups_concrete(
+        &mut self,
+        input: &T,
+        builder: &mut T::Builder,
+        groups: &EqGroups,
+    ) -> Result<()> {
+        let mut groups_iter = groups.0.iter().peekable();
+        let mut cur = self.group_value.as_ref().map(|x| x.as_scalar_ref());
+        for (i, v) in input.iter().enumerate() {
+            if groups_iter.peek() == Some(&&i) {
+                groups_iter.next();
+                ensure!(self.ongoing);
+                builder.append(cur)?;
+            }
+            self.ongoing = true;
+            cur = v;
+        }
+        self.group_value = cur.map(|x| x.to_owned_scalar());
+        Ok(())
+    }
+    pub fn output_concrete(&self, builder: &mut T::Builder) -> Result<()> {
+        builder.append(self.group_value.as_ref().map(|x| x.as_scalar_ref()))
+    }
+}
+
+macro_rules! impl_sorted_grouper {
+    ($input:ty, $input_variant:ident) => {
+        impl SortedGrouper for GeneralSortedGrouper<$input> {
+            fn split_groups(&self, input: &ArrayImpl) -> Result<EqGroups> {
+                if let ArrayImpl::$input_variant(i) = input {
+                    self.split_groups_concrete(i)
+                } else {
+                    Err(ErrorCode::InternalError(format!(
+                        "Input fail to match {}.",
+                        stringify!($input_variant)
+                    ))
+                    .into())
+                }
+            }
+            fn output(&self, builder: &mut ArrayBuilderImpl) -> Result<()> {
+                if let ArrayBuilderImpl::$input_variant(b) = builder {
+                    self.output_concrete(b)
+                } else {
+                    Err(ErrorCode::InternalError(format!(
+                        "Builder fail to match {}.",
+                        stringify!($input_variant)
+                    ))
+                    .into())
+                }
+            }
+            fn update_and_output_with_sorted_groups(
+                &mut self,
+                input: &ArrayImpl,
+                builder: &mut ArrayBuilderImpl,
+                groups: &EqGroups,
+            ) -> Result<()> {
+                if let (ArrayImpl::$input_variant(i), ArrayBuilderImpl::$input_variant(b)) =
+                    (input, builder)
+                {
+                    self.update_and_output_with_sorted_groups_concrete(i, b, groups)
+                } else {
+                    Err(ErrorCode::InternalError(format!(
+                        "Input fail to match {} or builder fail to match {}.",
+                        stringify!($input_variant),
+                        stringify!($input_variant)
+                    ))
+                    .into())
+                }
+            }
+        }
+    };
+}
+impl_sorted_grouper! { I16Array, Int16 }
+impl_sorted_grouper! { I32Array, Int32 }
+impl_sorted_grouper! { I64Array, Int64 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use risingwave_pb::data::data_type::TypeName;
+    use risingwave_pb::data::DataType;
+    use risingwave_pb::expr::agg_call::Type;
+    use risingwave_pb::expr::AggCall;
+
+    use super::*;
+    use crate::array::column::Column;
+    use crate::vector_op::agg::general_agg::{sum, GeneralAgg};
+    use crate::vector_op::agg::AggStateFactory;
+
+    #[test]
+    fn group_int32() -> Result<()> {
+        let mut g = GeneralSortedGrouper::<I32Array> {
+            ongoing: false,
+            group_value: None,
+        };
+        let mut builder = I32ArrayBuilder::new(0)?;
+
+        let input = I32Array::from_slice(&[Some(1), Some(1), Some(3)]).unwrap();
+        let eq = g.split_groups_concrete(&input)?;
+        g.update_and_output_with_sorted_groups_concrete(&input, &mut builder, &eq)?;
+        assert_eq!(eq.0, vec![2]);
+        let input = I32Array::from_slice(&[Some(3), Some(4), Some(4)]).unwrap();
+        let eq = g.split_groups_concrete(&input)?;
+        g.update_and_output_with_sorted_groups_concrete(&input, &mut builder, &eq)?;
+        assert_eq!(eq.0, vec![1]);
+
+        g.output_concrete(&mut builder)?;
+        assert_eq!(
+            builder.finish()?.iter().collect::<Vec<_>>(),
+            vec![Some(1), Some(3), Some(4)]
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn group_intersect() {
+        assert_eq!(
+            EqGroups::intersect(&[EqGroups(vec![0, 2, 4]), EqGroups(vec![1, 2, 5]),]).0,
+            vec![0, 1, 2, 4, 5]
+        );
+    }
+
+    #[test]
+    fn vec_agg_group() -> Result<()> {
+        let mut g0 = GeneralSortedGrouper::<I32Array> {
+            ongoing: false,
+            group_value: None,
+        };
+        let mut g0_builder = I32ArrayBuilder::new(0)?;
+        let mut g1 = GeneralSortedGrouper::<I32Array> {
+            ongoing: false,
+            group_value: None,
+        };
+        let mut g1_builder = I32ArrayBuilder::new(0)?;
+        let mut a = GeneralAgg::<I32Array, _, I64Array>::new(DataTypeKind::Int64, 0, sum);
+        let mut a_builder = I64ArrayBuilder::new(0)?;
+
+        let g0_input = I32Array::from_slice(&[Some(1), Some(1), Some(3)]).unwrap();
+        let eq0 = g0.split_groups_concrete(&g0_input)?;
+        let g1_input = I32Array::from_slice(&[Some(7), Some(8), Some(8)]).unwrap();
+        let eq1 = g1.split_groups_concrete(&g1_input)?;
+        let eq = EqGroups::intersect(&[eq0, eq1]);
+        g0.update_and_output_with_sorted_groups_concrete(&g0_input, &mut g0_builder, &eq)?;
+        g1.update_and_output_with_sorted_groups_concrete(&g1_input, &mut g1_builder, &eq)?;
+        let a_input = I32Array::from_slice(&[Some(1), Some(2), Some(3)]).unwrap();
+        a.update_and_output_with_sorted_groups_concrete(&a_input, &mut a_builder, &eq)?;
+
+        let g0_input = I32Array::from_slice(&[Some(3), Some(4), Some(4)]).unwrap();
+        let eq0 = g0.split_groups_concrete(&g0_input)?;
+        let g1_input = I32Array::from_slice(&[Some(8), Some(8), Some(8)]).unwrap();
+        let eq1 = g1.split_groups_concrete(&g1_input)?;
+        let eq = EqGroups::intersect(&[eq0, eq1]);
+        g0.update_and_output_with_sorted_groups_concrete(&g0_input, &mut g0_builder, &eq)?;
+        g1.update_and_output_with_sorted_groups_concrete(&g1_input, &mut g1_builder, &eq)?;
+        let a_input = I32Array::from_slice(&[Some(1), Some(2), Some(3)]).unwrap();
+        a.update_and_output_with_sorted_groups_concrete(&a_input, &mut a_builder, &eq)?;
+
+        g0.output_concrete(&mut g0_builder)?;
+        g1.output_concrete(&mut g1_builder)?;
+        a.output_concrete(&mut a_builder)?;
+        assert_eq!(
+            g0_builder.finish()?.iter().collect::<Vec<_>>(),
+            vec![Some(1), Some(1), Some(3), Some(4)]
+        );
+        assert_eq!(
+            g1_builder.finish()?.iter().collect::<Vec<_>>(),
+            vec![Some(7), Some(8), Some(8), Some(8)]
+        );
+        assert_eq!(
+            a_builder.finish()?.iter().collect::<Vec<_>>(),
+            vec![Some(1), Some(2), Some(4), Some(5)]
+        );
+        Ok(())
+    }
+
+    #[test]
+    fn vec_count_star() {
+        let mut g0 = GeneralSortedGrouper::<I32Array> {
+            ongoing: false,
+            group_value: None,
+        };
+        let mut g0_builder = I32ArrayBuilder::new(0).unwrap();
+        let prost = AggCall {
+            r#type: Type::Count as i32,
+            args: vec![],
+            return_type: Some(DataType {
+                type_name: TypeName::Int64 as i32,
+                ..Default::default()
+            }),
+        };
+        let mut a = AggStateFactory::new(&prost)
+            .unwrap()
+            .create_agg_state()
+            .unwrap();
+        let mut a_builder = a.return_type().create_array_builder(0).unwrap();
+
+        let input = I32Array::from_slice(&[Some(1), Some(1), Some(3)]).unwrap();
+        let eq = g0.split_groups_concrete(&input).unwrap();
+        g0.update_and_output_with_sorted_groups_concrete(&input, &mut g0_builder, &eq)
+            .unwrap();
+        a.update_and_output_with_sorted_groups(
+            &DataChunk::builder()
+                .columns(vec![Column::new(Arc::new(input.into()))])
+                .build(),
+            &mut a_builder,
+            &eq,
+        )
+        .unwrap();
+
+        let input = I32Array::from_slice(&[Some(3), Some(3), Some(3)]).unwrap();
+        let eq = g0.split_groups_concrete(&input).unwrap();
+        g0.update_and_output_with_sorted_groups_concrete(&input, &mut g0_builder, &eq)
+            .unwrap();
+        a.update_and_output_with_sorted_groups(
+            &DataChunk::builder()
+                .columns(vec![Column::new(Arc::new(input.into()))])
+                .build(),
+            &mut a_builder,
+            &eq,
+        )
+        .unwrap();
+
+        let input = I32Array::from_slice(&[Some(3), Some(4), Some(4)]).unwrap();
+        let eq = g0.split_groups_concrete(&input).unwrap();
+        g0.update_and_output_with_sorted_groups_concrete(&input, &mut g0_builder, &eq)
+            .unwrap();
+        a.update_and_output_with_sorted_groups(
+            &DataChunk::builder()
+                .columns(vec![Column::new(Arc::new(input.into()))])
+                .build(),
+            &mut a_builder,
+            &eq,
+        )
+        .unwrap();
+
+        g0.output_concrete(&mut g0_builder).unwrap();
+        a.output(&mut a_builder).unwrap();
+        assert_eq!(
+            g0_builder.finish().unwrap().iter().collect::<Vec<_>>(),
+            vec![Some(1), Some(3), Some(4)]
+        );
+        assert_eq!(
+            a_builder
+                .finish()
+                .unwrap()
+                .as_int64()
+                .iter()
+                .collect::<Vec<_>>(),
+            vec![Some(2), Some(5), Some(2)]
+        );
+    }
+}
