+++ rust/common/src/vector_op/agg/mod.rs
@@ -1,0 +1,7 @@
+mod aggregator;
+mod count_star;
+mod general_agg;
+mod general_sorted_grouper;
+
+pub use aggregator::{AggStateFactory, BoxedAggState};
+pub use general_sorted_grouper::{create_sorted_grouper, BoxedSortedGrouper, EqGroups};
